chore: 로또 기능을 어플리케이션으로 분리
